requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
smolmodels==0.1.0
numpy==1.26.3
pandas==2.2.0
jinja2==3.1.3
pytest==7.4.4
httpx==0.26.0
python-dotenv==1.0.1
aiofiles==23.2.1 